ai-module/requirements.txt
flask
flask-cors
psycopg2-binary
ollama
requests
